refactor: break up open ai into multiple files
